Split main() into a separate module
This facilitates reuse of modules that depend on the global variables
that were previously coupled in the main module.
